Closes #9: Updated RSA class based on help from Sebastian Ramacher
After posting about this bug in the PyCrypto mailing list, I got a really
helpful response from a guy named Sebastian Ramacher, which helped me
understand RSA better and provide a better implementation for RSA, based
on the internal implementation of PKCS1_OAEP in that library. PCKS1 is
now used as the cipher for the EJTP encryptor.

EJTP RSA no longer tries to support "decrypting" plaintext, which was
always a bit of an ill-thought-out hack anyways. Bad input data, bad
length, if it ain't valid it's a ValueError. This is more sane, less
acrobatic, and so on, but it also forces a redesign of part of the
EJTP protocol. That will be a different ticket though.

Once the RSA test was modified not to try to decrypt plaintext data,
which we no longer try to support as per the previous paragraph, the
test worked consistently. Once the protocol is modified to use real
signatures, I think I'll be able to confidently support RSA crypto
over the wire in EJTP projects. HUGE STEP FORWARD.