Made AES cipher object transient and pw hash a property.
I did this as the PyCrypto AES object doesn't appreciate being
used to encrypt and then decrypt; for whatever reason, it is a
stateful object and should be re-instantiated before each use.

Also, "hash" is a reserved word in python, so I changed it to
"pwhash", and made it an object property to work around having
the cipher as a throwaway object.