Modified occurrences of str and unicode to make it work with python2.x and 3.x
Since unicode from python2.x is (almost) equal with str from python3.x and
bytes works different in each version it's necessary to check the version
before working with this classes.
I added the module util.py2and3. It contains functions that make it easier to
work with both python 2.x and 3.x.